Eh, just prove the copies of Sam's progress theorem.
I think Ornela et al.'s notion of progress follows from the one we've
defined, but they introduce a fair bit of extra work to get there.  In
essence, as I've understood it, it amounts to saying that in an open
system, as long as you plug in suitably typed additional channels, you
get to a state without blocked processes.  This would amount to the
contextual notion of deadlock freedom we started with.