First shot at typing system for buffers.
This suffers from continued conflation of queues, vectors, and cons
lists.  Also, I removed the uses of closures---I think given the PLACES
page limit, and that there's no asynchronous approach to CP, that's not
the primary issue to focus on.  We can add them back later if they're
important/we have space.